+++ internal/gopher-lua/_state.go
@@ -3,7 +3,6 @@ package lua
 import (
 	"context"
 	"fmt"
-	"github.com/heroiclabs/nakama/v3/internal/gopher-lua/parse"
 	"io"
 	"math"
 	"os"
@@ -12,6 +11,8 @@ import (
 	"sync"
 	"sync/atomic"
 	"time"
+
+	"github.com/heroiclabs/nakama/v3/internal/gopher-lua/parse"
 )
 
 const MultRet = -1
@@ -1216,6 +1217,10 @@ func NewState(opts ...Options) *LState {
 	return ls
 }
 
+func (ls *LState) IsClosed() bool {
+	return ls.stack == nil
+}
+
 func (ls *LState) Close() {
 	atomic.AddInt32(&ls.stop, 1)
 	for _, file := range ls.G.tempFiles {
@@ -2020,7 +2025,7 @@ func (ls *LState) SetMx(mx int) {
 	go func() {
 		limit := uint64(mx * 1024 * 1024) //MB
 		var s runtime.MemStats
-		for ls.stop == 0 {
+		for atomic.LoadInt32(&ls.stop) == 0 {
 			runtime.ReadMemStats(&s)
 			if s.Alloc >= limit {
 				fmt.Println("out of memory")