+++ internal/gopher-lua/oslib.go
@@ -15,9 +15,24 @@ func init() {
 
 func getIntField(L *LState, tb *LTable, key string, v int) int {
 	ret := tb.RawGetString(key)
-	if ln, ok := ret.(LNumber); ok {
-		return int(ln)
+
+	switch lv := ret.(type) {
+	case LNumber:
+		return int(lv)
+	case LString:
+		slv := string(lv)
+		slv = strings.TrimLeft(slv, " ")
+		if strings.HasPrefix(slv, "0") && !strings.HasPrefix(slv, "0x") && !strings.HasPrefix(slv, "0X") {
+			//Standard lua interpreter only support decimal and hexadecimal
+			slv = strings.TrimLeft(slv, "0")
+		}
+		if num, err := parseNumber(slv); err == nil {
+			return int(num)
+		}
+	default:
+		return v
 	}
+
 	return v
 }